Refactor API checkAccess
